test: move PharmCATTest and rework how tests save files

+++ src/test/java/org/pharmgkb/pharmcat/TestUtils.java
@@ -14,10 +14,30 @@ import org.junit.jupiter.api.TestInfo;
  * @author Mark Woon
  */
 public class TestUtils {
-  public static final Path TEST_OUTPUT_DIR = getTestOutputDir();
+  private static Path s_testOutputDir = getDefaultTestOutputDir();
+  private static boolean s_saveTestOutput = "true".equals(System.getenv("PHARMCAT_SAVE_TEST_OUTPUT"));
+
 
   private TestUtils() {
   }
+
+  public static void setTestOutputDir(Path outputDir) {
+    s_testOutputDir = outputDir;
+  }
+
+  public static Path getTestOutputDir() {
+    return s_testOutputDir;
+  }
+
+
+  public static boolean isSaveTestOutput() {
+    return s_saveTestOutput;
+  }
+
+  public static void setSaveTestOutput(boolean saveTestOutput) {
+    s_saveTestOutput = saveTestOutput;
+  }
+
 
 
   public static String getTestName(TestInfo testInfo) {
@@ -46,14 +66,17 @@ public class TestUtils {
   }
 
 
-  private static Path getTestOutputDir() {
+
+  private static Path getDefaultTestOutputDir() {
     Path outputDir;
     if (System.getProperty("PHARMCAT_TEST_DIR") != null) {
       outputDir = Paths.get(System.getProperty("PHARMCAT_TEST_DIR"));
     } else {
       outputDir = Paths.get("out");
       if (!outputDir.isAbsolute()) {
-        outputDir = Paths.get(System.getProperty("user.dir")).resolve("out");
+        outputDir = Paths.get(System.getProperty("user.dir")).resolve("tmp/test_output");
+      } else {
+        outputDir = outputDir.resolve("test_output");
       }
     }
     if (!Files.isDirectory(outputDir)) {
@@ -67,14 +90,37 @@ public class TestUtils {
   }
 
 
+  /**
+   * Gets the output directory for the given test.
+   * Directory is guaranteed to exist.
+   *
+   * @param deleteIfExist if directory exists, it will be deleted and re-created
+   */
   public static Path getTestOutputDir(TestInfo testInfo, boolean deleteIfExist) throws IOException {
     Path classOutputDir;
     if (testInfo.getTestClass().isPresent()) {
-      classOutputDir = TEST_OUTPUT_DIR.resolve(testInfo.getTestClass().get().getSimpleName());
+      classOutputDir = s_testOutputDir.resolve(testInfo.getTestClass().get().getSimpleName());
     } else {
-      classOutputDir = TEST_OUTPUT_DIR;
+      classOutputDir = s_testOutputDir;
     }
     Path dir = classOutputDir.resolve(getTestName(testInfo));
+    if (Files.exists(dir)) {
+      if (Files.isDirectory(dir)) {
+        if (deleteIfExist) {
+          FileUtils.deleteDirectory(dir.toFile());
+        }
+      } else {
+        throw new RuntimeException("Not a directory: " + dir);
+      }
+    }
+    if (!Files.exists(dir)) {
+      Files.createDirectories(dir);
+    }
+    return dir;
+  }
+
+  public static Path getTestOutputDir(Class testClass, boolean deleteIfExist) throws IOException {
+    Path dir = s_testOutputDir.resolve(testClass.getSimpleName());
     if (Files.exists(dir)) {
       if (Files.isDirectory(dir)) {
         if (deleteIfExist) {
@@ -87,10 +133,11 @@ public class TestUtils {
     return dir;
   }
 
-  public static Path createTempFile(String prefix, String suffix) throws IOException {
-    return createTempFile(TEST_OUTPUT_DIR, prefix, suffix);
-  }
 
+  /**
+   * Creates a temporary test file based on {@code testInfo}, with the specified suffix.
+   * This file name will be change from one test to the next.
+   */
   public static Path createTempFile(TestInfo testInfo, String suffix) throws IOException {
     return createTempFile(getTestOutputDir(testInfo, false), getTestName(testInfo), suffix);
   }
@@ -100,11 +147,49 @@ public class TestUtils {
       Files.createDirectories(dir);
     }
     Path file = Files.createTempFile(dir, prefix, suffix);
-    file.toFile().deleteOnExit();
+    if (!s_saveTestOutput) {
+      file.toFile().deleteOnExit();
+    }
     return file;
   }
 
   public static Path createTempDirectory(String prefix) throws IOException {
-    return Files.createTempDirectory(TEST_OUTPUT_DIR, prefix);
+    return Files.createTempDirectory(s_testOutputDir, prefix);
+  }
+
+
+  /**
+   * Creates a test file based on {@code testInfo}, with the specified suffix.
+   * This file name will be constant from one test to the next.
+   */
+  public static Path createTestFile(TestInfo testInfo, String suffix) throws IOException {
+    return createTestFile(getTestOutputDir(testInfo, false), getTestName(testInfo) + suffix);
+  }
+
+  public static Path createTestFile(Class testClass, String filename) throws IOException {
+    return createTestFile(getTestOutputDir(testClass, false), filename);
+  }
+
+  /**
+   * Creates a test file with the given filename.
+   * This file name will be constant from one test to the next.
+   */
+  private static Path createTestFile(Path dir, String filename) throws IOException {
+    if (!Files.isDirectory(dir)) {
+      Files.createDirectories(dir);
+    }
+    Path file = dir.resolve(filename);
+    if (!s_saveTestOutput) {
+      file.toFile().deleteOnExit();
+    }
+    return file;
+  }
+
+  public static void deleteTestFiles(Path... files) throws IOException {
+    if (!isSaveTestOutput()) {
+      for (Path file : files) {
+        Files.deleteIfExists(file);
+      }
+    }
   }
 }
